% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_sse.R
\name{Regress_SSE}
\alias{Regress_SSE}
\title{Regress_SSE}
\usage{
Regress_SSE(
  a_data,
  nseq_SampleSizeToTest,
  c_learner,
  c_measurment,
  n_bootstraps,
  c_RegressionTarget
)
}
\arguments{
\item{a_data}{A data set object, containing oberservations of the parameters}

\item{nseq_SampleSizeToTest}{A sequence of integers, defining the numbers of sample sizes, which should be tested}

\item{c_learner}{The name of a regression learner from the mlr3 package}

\item{c_measurment}{The name of a measure for regressiors from the mlr3 package}

\item{n_bootstrap}{The number of bootstraps, how often the learning task should be repeated}

\item{c_classificationTarget}{A parameter of the data object, which should be used as target for the regression}
}
\value{
The output can be used for plotting in the *plotting function
}
\description{
This function provides a simulation for the behavior of a defined measure of a learners performance with increasing numbers of cases
}
