% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthDataGen.R
\name{SynthDataGen}
\alias{SynthDataGen}
\title{SynthDataGen}
\usage{
SynthDataGen(a_OriginalData, i_OutputData, c_method = "amelia")
}
\arguments{
\item{a_OriginalData}{The data, which is available and should be used for the data generation, in form of a data frame}

\item{n_OutputData}{A integer indicating the number of cases, which should be there at the and of generation.}

\item{c_methode}{The method, which should be used for data generation, currently only the amelia algorithm is available}
}
\value{
The output is a array, which contains the original and generated data
}
\description{
This function provides a data generation, while using existing data and imputation algorithms.
}
